backend/requirements.txt
# --- Core Framework ---
fastapi==0.111.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.9.1

# --- Database & Migrations ---
SQLAlchemy==2.0.41
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3
aiosqlite==0.20.0

# --- Security & Auth ---
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
email_validator==2.2.0

# --- API Clients & HTTP ---
google-generativeai==0.8.5
googlemaps==4.10.0
httpx==0.27.2
wikipedia==1.4.0
requests==2.32.4
polyline==2.0.0

# --- Caching & Rate Limiting ---
fastapi-cache2[redis]
slowapi

# --- Utilities ---
opening-hours-py==1.1.3

# --- Testing ---
pytest
pytest-asyncio

# --- Production Server ---
gunicorn==22.0.0

# This is often needed for local PostgreSQL setups on some systems
psycopg2-binary==2.9.9